+++ planner.h
@@ -34,10 +34,10 @@ typedef struct {
   int32_t  step_event_count;          // The number of step events required to complete this block
 
   // Fields used by the motion planner to manage acceleration
-  double nominal_speed;               // The nominal speed for this block in mm/min  
-  double entry_speed;                 // Entry speed at previous-current block junction in mm/min
-  double max_entry_speed;             // Maximum allowable junction entry speed in mm/min
-  double millimeters;                 // The total travel of this block in mm
+  float nominal_speed;               // The nominal speed for this block in mm/min  
+  float entry_speed;                 // Entry speed at previous-current block junction in mm/min
+  float max_entry_speed;             // Maximum allowable junction entry speed in mm/min
+  float millimeters;                 // The total travel of this block in mm
   uint8_t recalculate_flag;           // Planner flag to recalculate trapezoids on entry junction
   uint8_t nominal_length_flag;        // Planner flag for nominal speed always reached
 
@@ -57,7 +57,7 @@ void plan_init();
 // Add a new linear movement to the buffer. x, y and z is the signed, absolute target position in 
 // millimaters. Feed rate specifies the speed of the motion. If feed rate is inverted, the feed
 // rate is taken to mean "frequency" and would complete the operation in 1/feed_rate minutes.
-void plan_buffer_line(double x, double y, double z, double feed_rate, uint8_t invert_feed_rate);
+void plan_buffer_line(float x, float y, float z, float feed_rate, uint8_t invert_feed_rate);
 
 // Called when the current block is no longer needed. Discards the block and makes the memory
 // availible for new blocks.
